Implement Friendship V2
Implemented the friendship v2 by creating a FriendRequest model to
track friendship and creating Friendship model to track friendship.
After a user accepts a request a transaction is commited to create a
friefriendship object to correspond to the request and mirror it. Then a
service object is called to destroy the friend request